Move model additions to their ORM specific file locations
The main model_additions.rb was getting cluttered with a bunch of
ORM tests. This can clean up a bit and only pull in the missing bits
and keep the common bits at a higher level, but since the file is
small now, it is just easier to duplicate the effort in each ORM
specific file.